# docs/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Varka Backend API
  version: 0.0.1
  description: REST API for authentication, projects, CIM/PIM/PSM models, transformations, artifacts, modeling metadata, and diagram layout.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:8080
    description: Local development
security:
  - AuthToken: []
...
paths:
...
  /api/chatbot/conversations:
    get:
      tags: [Assistant]
      summary: List recent assistant conversations
      responses:
        "200":
          description: Recent conversations
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/chatbot/sessions:
    post:
      tags: [Assistant]
      summary: Create assistant session
      responses:
        "200":
          description: Created assistant session
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/chatbot/sessions/{sessionId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
    delete:
      tags: [Assistant]
      summary: Clear assistant session memory
      responses:
        "200":
          description: Session cleared
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/chatbot/sessions/{sessionId}/messages:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Assistant]
      summary: Send assistant message
      responses:
        "200":
          description: Assistant response accepted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/chatbot/sessions/{sessionId}/cancel:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Assistant]
      summary: Cancel active assistant turn
      responses:
        "200":
          description: Cancellation accepted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/chatbot/sessions/{sessionId}/attachments:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Assistant]
      summary: Upload assistant attachment
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Uploaded attachment metadata
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /api/chatbot/sessions/{sessionId}/events:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Assistant]
      summary: Open assistant SSE event stream
      responses:
        "200":
          description: Server-sent event stream
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/chatbot/sessions/{sessionId}/thread:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Assistant]
      summary: Get durable assistant thread history
      responses:
        "200":
          description: Assistant thread snapshot
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/chatbot/sessions/{sessionId}/proposals/{proposalId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Assistant]
      summary: Get assistant proposal
      responses:
        "200":
          description: Assistant proposal
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/chatbot/sessions/{sessionId}/proposals/{proposalId}/undo:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Assistant]
      summary: Undo applied assistant proposal
      responses:
        "200":
          description: Proposal undone
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    AuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
...
  responses:
...
    Unauthorized:
      description: Missing, invalid, or expired auth session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
    PayloadTooLarge:
      description: Upload or generated payload exceeds configured limits
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    UnsupportedMediaType:
      description: Uploaded media type or file format is not supported
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
  schemas:
    ApiErrorResponse:
      type: object
      required: [message, status, timestamp, issues, errorId]
      properties:
        message:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        issues:
          type: array
          items: {}
        errorId:
          type: string
          description: Correlation id for support and log lookup
...